Ingestion/Adzuna/requirements.txt
python-dotenv==1.1.0
requests==2.32.3
#jupyter
notebook
ipykernel

pandas
